Add OpenSearch store
The newer Elasticsearch client libraries are not compatible with
OpenSearch.
