+++ constraints.txt
@@ -97,6 +97,7 @@ pysaml2==7.1.2
 python-dateutil==2.8.2
 python-ldap==3.4.2
 pytz==2022.1
+opensearch-py==2.0.0
 redis==4.3.4
 requests-aws4auth==1.1.2
 requests-oauthlib==1.3.1
